WASM_X86: Combine X86Generator and X86Visitor Classes
